Make better use of the Arrow Table feature
`bind_rows` isn't implemented for Arrow Table but doing reduce over `union_all` achieves the same result. It is much faster to do this to an Arrow Table, rather than tibbles.

I have also added the option for the user to return the object as an Arrow Table, although a tibble is still the default.