Update render-README.yaml
Made changes to use stefanzweifel workflow for commit to production and branches rather than the former logiv that uses peter-evans/create-pull-request@v7 for commits and PR on production and stefanzweifel/git-auto-commit-action@v6 for other branches. 

This is to fix the error 128 detached HEAD that occurs when we merge other branches to production.